tests/wpt/tests/html/canvas/offscreen/layers/2d.layer.invalid-calls.endLayer.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.layer.invalid-calls.endLayer</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.layer.invalid-calls.endLayer</h1>
<p class="desc">Raises exception on lone endLayer calls.</p>


<script>
test(t => {
  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  assert_throws_dom("INVALID_STATE_ERR", function() {
    ctx.endLayer();
  });

}, "Raises exception on lone endLayer calls.");
</script>
